Rename furnitureLayouts to furnitureArrangements
