+++ dotcom-rendering/src/layouts/lib/furnitureArrangements.ts
